# api/openapi.yaml
openapi: 3.0.3
info:
  title: iRODS REST API
  version: 0.1.0
  description: |
    REST-ful iRODS running on GoLang.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/principal:
    get:
      operationId: getPrincipals
      tags: [user]
      summary: Search iRODS users and groups
      description: |
        Searches user and group principals by name prefix for combined selector
        UIs. Use the separate `/api/v1/user` and `/api/v1/usergroup` prefix
        searches when a combined result set is not needed.
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - in: query
          name: zone
          required: false
          schema:
            type: string
          description: iRODS zone to search. Defaults to the configured iRODS zone.
        - in: query
          name: query
          required: true
          schema:
            type: string
            minLength: 3
          description: Principal name prefix to search.
        - in: query
          name: kind
          required: false
          schema:
            type: array
            items:
              type: string
              enum: [user, group]
          style: form
          explode: true
          description: Optional principal kind filter. May be repeated.
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
          description: Maximum result count.
      responses:
        "200":
          description: Matching principals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrincipalSearchResponse"
        "400":
          description: Invalid search option
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Access denied while searching principals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/user:
    get:
      operationId: getUsers
      tags: [user]
      summary: List iRODS users
      description: |
        Lists visible iRODS user principals in the configured zone. When `prefix`
        is supplied for autocomplete-style search, it must be at least 3 characters.
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - in: query
          name: zone
          required: false
          schema:
            type: string
          description: iRODS zone to search. Defaults to the configured iRODS zone.
        - in: query
          name: type
          required: false
          schema:
            type: string
            enum: [rodsuser, rodsadmin]
          description: Optional user type filter.
        - in: query
          name: prefix
          required: false
          schema:
            type: string
            minLength: 3
          description: Optional username prefix filter for autocomplete.
      responses:
        "200":
          description: Users visible in the zone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserCollectionResponse"
        "400":
          description: Invalid user list option
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Access denied while listing users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: postUser
      tags: [user]
      summary: Create one iRODS user
      description: |
        Requires the authenticated iRODS user to have type `rodsadmin` or `groupadmin`.
        Synchronization clients may use `reconcile=true` to make creation repeatable
        when the existing principal is already the requested user type. The operation
        remains a conflict when the existing principal is a group or when the existing
        user type differs from the requested type.
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - in: query
          name: zone
          required: false
          schema:
            type: string
          description: iRODS zone. Defaults to the configured iRODS zone.
        - in: query
          name: reconcile
          required: false
          schema:
            type: boolean
            default: false
          description: |
            Desired-state sync mode. Treat an existing user with the requested type
            as success instead of returning 409. A type mismatch or non-user
            principal with the same name remains a conflict. If a password is supplied
            for an existing matching user, the password is applied.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreateRequest"
      responses:
        "200":
          description: Desired user state satisfied when reconciliation mode is used
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Invalid create request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Authenticated iRODS user is not rodsadmin or groupadmin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/user/me:
    get:
      operationId: getCurrentUserMembership
      tags: [user]
      summary: Fetch the authenticated iRODS user and groups
      description: |
        Returns the effective authenticated iRODS user, group memberships, and
        computed administration flags for Starbase user/group UI decisions.
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - in: query
          name: zone
          required: false
          schema:
            type: string
          description: iRODS zone. Defaults to the configured iRODS zone.
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
          description: Maximum group count.
      responses:
        "200":
          description: Current iRODS user membership and administration flags
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUserMembershipResponse"
        "400":
          description: Invalid current-user option
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Access denied while fetching current user membership
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/user/membership-summary:
    get:
      operationId: getUserMembershipSummaries
      tags: [user]
      summary: List users with group membership summaries
      description: |
        Returns user rows with group names for list-scale administration UIs.
        This avoids broad client-side joins or repeated per-user group lookups.
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - in: query
          name: zone
          required: false
          schema:
            type: string
          description: iRODS zone to search. Defaults to the configured iRODS zone.
        - in: query
          name: type
          required: false
          schema:
            type: string
            enum: [rodsuser, rodsadmin]
          description: Optional user type filter.
        - in: query
          name: prefix
          required: false
          schema:
            type: string
            minLength: 3
          description: Optional username prefix filter.
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
          description: Maximum user count.
      responses:
        "200":
          description: Users with membership summaries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserMembershipSummaryCollectionResponse"
        "400":
          description: Invalid summary option
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Access denied while listing user memberships
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v1/user/{user_name}/usergroup:
    get:
      operationId: getUserGroupsForUser
      tags: [user]
      summary: List groups containing one user
      description: |
        Returns the iRODS groups containing the selected user. This is the
        reverse-membership lookup for user detail and administration UIs.
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - in: path
          name: user_name
          required: true
          schema:
            type: string
          description: iRODS username.
        - in: query
          name: zone
          required: false
          schema:
            type: string
          description: iRODS zone to search. Defaults to the configured iRODS zone.
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
          description: Maximum group count.
      responses:
        "200":
          description: Groups containing the user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserGroupRefCollectionResponse"
        "400":
          description: Invalid reverse membership option
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Access denied while listing user groups
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v1/usergroup:
    get:
      operationId: getUserGroups
      tags: [user]
      summary: List iRODS groups
      description: |
        Lists visible iRODS groups in the configured zone. When `prefix`
        is supplied for autocomplete-style search, it must be at least 3 characters.
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - in: query
          name: zone
          required: false
          schema:
            type: string
          description: iRODS zone to search. Defaults to the configured iRODS zone.
        - in: query
          name: prefix
          required: false
          schema:
            type: string
            minLength: 3
          description: Optional group-name prefix filter for autocomplete.
      responses:
        "200":
          description: Groups visible in the zone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserGroupCollectionResponse"
        "400":
          description: Invalid group list option
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Access denied while listing groups
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: postUserGroup
      tags: [user]
      summary: Create one iRODS group
      description: |
        Generic REST surface for creating iRODS groups. Workflows that
        need to provision iRODS groups should call this route rather than an
        `/api/v1/ext` route.

        Requires the authenticated iRODS user to have type `rodsadmin` or `groupadmin`.
        Synchronization clients may use `reconcile=true` to make group creation
        repeatable when the existing principal is already an iRODS group. Created
        groups are marked with `iRODS:USER_SYNCH:MANAGED=true`.
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - in: query
          name: zone
          required: false
          schema:
            type: string
          description: iRODS zone. Defaults to the configured iRODS zone.
        - in: query
          name: reconcile
          required: false
          schema:
            type: boolean
            default: false
          description: |
            Desired-state sync mode. Treat an existing group as success instead of
            returning 409. A same-name non-group principal remains a conflict.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserGroupCreateRequest"
      responses:
        "200":
          description: Desired group state satisfied when reconciliation mode is used
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserGroupResponse"
        "201":
          description: Group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserGroupResponse"
        "400":
          description: Invalid create request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Group already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Authenticated iRODS user is not rodsadmin or groupadmin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/usergroup/summary:
    get:
      operationId: getUserGroupSummaries
      tags: [user]
      summary: List iRODS groups with member counts
      description: |
        Returns group rows with member counts for list-scale administration UIs.
        This avoids repeated per-group detail fetches from the browser.
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - in: query
          name: zone
          required: false
          schema:
            type: string
          description: iRODS zone to search. Defaults to the configured iRODS zone.
        - in: query
          name: prefix
          required: false
          schema:
            type: string
            minLength: 3
          description: Optional group-name prefix filter.
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
          description: Maximum group count.
      responses:
        "200":
          description: Groups with member counts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserGroupSummaryCollectionResponse"
        "400":
          description: Invalid group summary option
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Access denied while listing group summaries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    BasicAuth:
      type: http
      scheme: basic
  schemas:
...
    ActionLink:
      type: object
      required: [href]
      properties:
        href:
          type: string
        method:
          type: string
...
    User:
      type: object
      required: [id, name, zone, type]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        zone:
          type: string
        type:
          type: string
          enum: [rodsuser, rodsadmin, groupadmin]
        links:
          $ref: "#/components/schemas/UserLinks"
    UserLinks:
      type: object
      properties:
        self:
          $ref: "#/components/schemas/ActionLink"
        update:
          $ref: "#/components/schemas/ActionLink"
        delete:
          $ref: "#/components/schemas/ActionLink"
        avus:
          $ref: "#/components/schemas/ActionLink"
        create_avu:
          $ref: "#/components/schemas/ActionLink"
    UserResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
    UserCollectionResponse:
      type: object
      required: [users, count, zone, links]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        count:
          type: integer
        zone:
          type: string
        type:
          type: string
          enum: [rodsuser, rodsadmin, groupadmin]
        prefix:
          type: string
        links:
          type: object
          properties:
            self:
              $ref: "#/components/schemas/ActionLink"
            create:
              $ref: "#/components/schemas/ActionLink"
    UserMembershipSummary:
      type: object
      required: [id, name, zone, type, groups]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        zone:
          type: string
        type:
          type: string
          enum: [rodsuser, rodsadmin, groupadmin]
        groups:
          type: array
          items:
            $ref: "#/components/schemas/UserGroupRef"
    CurrentUserMembership:
      type: object
      required:
        - user
        - groups
        - is_rodsadmin
        - is_groupadmin
        - can_administer_users_and_groups
      properties:
        user:
          $ref: "#/components/schemas/User"
        groups:
          type: array
          items:
            $ref: "#/components/schemas/UserGroupRef"
        is_rodsadmin:
          type: boolean
        is_groupadmin:
          type: boolean
        can_administer_users_and_groups:
          type: boolean
    CurrentUserMembershipResponse:
      type: object
      required: [current_user, zone]
      properties:
        current_user:
          $ref: "#/components/schemas/CurrentUserMembership"
        zone:
          type: string
        limit:
          type: integer
        links:
          type: object
          properties:
            self:
              $ref: "#/components/schemas/ActionLink"
            groups:
              $ref: "#/components/schemas/ActionLink"
    UserMembershipSummaryCollectionResponse:
      type: object
      required: [users, count, zone]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserMembershipSummary"
        count:
          type: integer
        zone:
          type: string
        type:
          type: string
          enum: [rodsuser, rodsadmin]
        prefix:
          type: string
        limit:
          type: integer
    UserCreateRequest:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
          description: iRODS username to create.
        type:
          type: string
          enum: [rodsuser, rodsadmin]
          description: iRODS user type for the new user.
        password:
          type: string
          format: password
          description: Optional initial iRODS user password.
...
    UserGroup:
      type: object
      required: [id, name, zone, type]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        zone:
          type: string
        type:
          type: string
          enum: [rodsgroup]
        members:
          type: array
          items:
            $ref: "#/components/schemas/UserGroupMember"
        links:
          $ref: "#/components/schemas/UserGroupLinks"
    UserGroupMember:
      type: object
      required: [id, name, zone, type]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        zone:
          type: string
        type:
          type: string
          enum: [rodsuser, rodsadmin, groupadmin]
        links:
          $ref: "#/components/schemas/UserGroupMemberLinks"
    UserGroupLinks:
      type: object
      properties:
        self:
          $ref: "#/components/schemas/ActionLink"
        delete:
          $ref: "#/components/schemas/ActionLink"
        add_member:
          $ref: "#/components/schemas/ActionLink"
        avus:
          $ref: "#/components/schemas/ActionLink"
        create_avu:
          $ref: "#/components/schemas/ActionLink"
    UserGroupMemberLinks:
      type: object
      properties:
        self:
          $ref: "#/components/schemas/ActionLink"
        remove_from_group:
          $ref: "#/components/schemas/ActionLink"
    UserGroupRef:
      type: object
      required: [id, name, zone, type]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        zone:
          type: string
        type:
          type: string
          enum: [rodsgroup]
    UserGroupRefCollectionResponse:
      type: object
      required: [groups, count, user_name, zone]
      properties:
        groups:
          type: array
          items:
            $ref: "#/components/schemas/UserGroupRef"
        count:
          type: integer
        user_name:
          type: string
        zone:
          type: string
        limit:
          type: integer
    UserGroupSummary:
      type: object
      required: [id, name, zone, type, member_count]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        zone:
          type: string
        type:
          type: string
          enum: [rodsgroup]
        member_count:
          type: integer
    UserGroupSummaryCollectionResponse:
      type: object
      required: [groups, count, zone]
      properties:
        groups:
          type: array
          items:
            $ref: "#/components/schemas/UserGroupSummary"
        count:
          type: integer
        zone:
          type: string
        prefix:
          type: string
        limit:
          type: integer
    UserGroupResponse:
      type: object
      required: [group]
      properties:
        group:
          $ref: "#/components/schemas/UserGroup"
    UserGroupCollectionResponse:
      type: object
      required: [groups, count, zone, links]
      properties:
        groups:
          type: array
          items:
            $ref: "#/components/schemas/UserGroup"
        count:
          type: integer
        zone:
          type: string
        prefix:
          type: string
        links:
          type: object
          properties:
            self:
              $ref: "#/components/schemas/ActionLink"
            create:
              $ref: "#/components/schemas/ActionLink"
    UserGroupCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          description: iRODS group name to create.
...
    PrincipalSearchResult:
      type: object
      required: [id, name, zone, type, kind, score]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        zone:
          type: string
        type:
          type: string
          enum: [rodsuser, rodsadmin, groupadmin, rodsgroup]
        kind:
          type: string
          enum: [user, group]
        score:
          type: integer
    PrincipalSearchResponse:
      type: object
      required: [principals, count, zone, query]
      properties:
        principals:
          type: array
          items:
            $ref: "#/components/schemas/PrincipalSearchResult"
        count:
          type: integer
        zone:
          type: string
        query:
          type: string
        kinds:
          type: array
          items:
            type: string
            enum: [user, group]
        limit:
          type: integer
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        fields:
          type: object
          description: Optional field-level validation messages keyed by request field name.
          additionalProperties:
            type: string